compatibility_update

+++ fractol/src/libft/ft_strcmp.c
@@ -6,7 +6,7 @@
 /*   By: tbrunier <tbrunier@student.42perpignan.fr> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/18 18:44:29 by tbrunier          #+#    #+#             */
-/*   Updated: 2023/10/18 18:44:29 by tbrunier         ###   ########.fr       */
+/*   Updated: 2024/09/18 17:18:40 by tbrunier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strcmp(const char *s1, const char *s2)
 {
-	size_t	i;
+	long long	i;
 
 	i = 0;
 	while (s1[i] == s2[i] && s1[i] != '\0' && s2[i] != '\0')
